GoogleCloudRunFunc/requirements.txt
google-cloud-aiplatform
opencv-python-headless
numpy
requests
functions-framework